move the ZFS identification parts to identify
